[refactor] ManageViewModel ChallengeFetchUsecase 적용
